% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dwnld.R
\name{data_html}
\alias{data_html}
\title{Work in progress copié de https://github.com/Pakillo/rmdTemplates/blob/master/R/embed_data_html.R}
\usage{
data_html(dataset, textHTML = "Click here to download the data")
}
\arguments{
\item{dataset}{data}

\item{textHTML}{text}
}
\value{

}
\description{
Work in progress copié de https://github.com/Pakillo/rmdTemplates/blob/master/R/embed_data_html.R
}
