% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitLayout.R
\name{I}
\alias{I}
\title{DigitLayout}
\usage{
I(x, digits = NULL, ...)
}
\arguments{
\item{x}{un vecteur numérique}

\item{digits}{nombre après la virgule}

\item{...}{,,,}
}
\value{
un vecteur numérique mis en forme selon les normes française
}
\description{
DigitLayout
}
\examples{
test = c(112582,4,524.218,67,.1,.14444,0)
I(test,digits = 2)
}
