% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{mode_stat}
\alias{mode_stat}
\title{Mode unique}
\usage{
mode_stat(x, na.rm = FALSE)
}
\arguments{
\item{x}{vecteur}

\item{na.rm}{logical, indique si les NA doivent etre pris en compte (FALSE par defaut)}
}
\value{
un character (si le mode correspond à plusieurs éléments, indique le premier)
}
\description{
Mode unique
}
\examples{
test = c(1,1,1,1,2,2,2,2,3,4,4,4,4)
mode_stat (test)
}
