% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{modes_stat}
\alias{modes_stat}
\title{Modes liste}
\usage{
modes_stat(x, na.rm = FALSE)
}
\arguments{
\item{x}{vecteur}

\item{na.rm}{logical, indique si les NA doivent etre pris en compte (FALSE par defaut)}
}
\value{
chaine de character (si le mode correspond à plusieurs éléments, les indique tous)
}
\description{
Modes liste
}
\examples{
test = c(1,1,1,1,2,2,2,2,3,4,4,4,4)
modes_stat(test)
}
